Borre duplicacion en ExternalizablePropertyDrawers
El codigo de ambos ExternalizableProperties era igual asi que cree una clase padre de la que heredan las otras dos y solo se marcan los Attributos hasta arriba de la clase para indicar qué clase deben dibujar
