+++ Editor/PropertyDrawers/ExternalizableProperty/ExternalizableProperty.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3d2e12079480f7a4a8e3e2becef0e22a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
